requirements.txt
pandas
tqdm
numpy
scipy
xarray
apscheduler
fastapi
uvicorn[standard]
google-cloud-storage
python-dotenv